+++ src/components/icon/icon.jsx
@@ -2,6 +2,9 @@ import React from 'react';
 import PureComponent from 'react-pure-render/component';
 import { elementType } from 'react-prop-types';
 import { IconStateless } from 'react-svg-sprite-icon';
+import classNames from 'classnames';
+import kebabCase from 'lodash.kebabcase';
+import assign from 'lodash.assign';
 import icons from './icons';
 
 class Icon extends PureComponent {
@@ -10,10 +13,49 @@ class Icon extends PureComponent {
     this.icons = icons;
   }
 
+  renderInline() {
+    const {
+      className,
+      style,
+      type,
+      fill,
+      stroke,
+      strokeWidth,
+      width,
+      height,
+      ...rest,
+    } = this.props;
+
+    const icon = this.icons[type]({ rootElement: 'svg', fill, stroke, strokeWidth });
+    const classes = classNames('icon', `icon--${kebabCase(type)}`, className);
+    const styles = assign({
+      display: 'inline-block',
+      width: `${(width || icon.info.width) / 10}rem`,
+      height: `${(height || icon.info.height) / 10}rem`,
+    }, style);
+
+    return (
+      <span
+        className={ classes }
+        style={ styles }
+        dangerouslySetInnerHTML={ { __html: icon.data } }
+        { ...rest }
+      />
+    );
+  }
+
   render() {
+    if (this.props.renderInline) {
+      return this.renderInline();
+    }
+
     return <IconStateless { ...this.props } name={ this.props.type } svg={ this.icons[this.props.type] } />;
   }
 }
+
+Icon.defaultProps = {
+  renderInline: false,
+};
 
 Icon.propTypes = {
   className: React.PropTypes.string,
@@ -28,6 +70,8 @@ Icon.propTypes = {
   /** Unitless pixel value */
   height: React.PropTypes.number,
   style: React.PropTypes.object,
+  /** Useful for server side rendering */
+  renderInline: React.PropTypes.bool,
   componentClass: elementType,
 };