+++ src/components/nav-bar/index.js
@@ -1,0 +1,1 @@
+export { default } from './nav-bar.jsx';