+++ src/index.js
@@ -8,8 +8,10 @@ import GraphTooltip from './components/graph-tooltip';
 import Icon from './components/icon/icon';
 import Input from './components/input/input';
 import LabeledValue from './components/labeled-value/labeled-value';
+import Legend from './components/legend';
 import Logo from './components/logo/logo';
 import Pane from './components/pane/pane';
+import RangeSelector from './components/range-selector/';
 import RatioBar from './components/ratio-bar';
 import Select from './components/select/select';
 import SparkGraph from './components/spark-graph';
@@ -35,8 +37,10 @@ export {
   Icon,
   Input,
   LabeledValue,
+  Legend,
   Logo,
   Pane,
+  RangeSelector,
   RatioBar,
   Select,
   SparkGraph,