+++ svgo.config.js
@@ -1,0 +1,10 @@
+module.exports = JSON.stringify({
+  plugins: [
+    { removeTitle: true },
+    { removeComments: true },
+    { removeDesc: true },
+    { removeDimensions: true },
+    { convertColors: { shorthex: true } },
+    { cleanupIDs: false },
+  ],
+});