+++ webpack.config.babel.js
@@ -4,10 +4,14 @@ import autoprefixer from 'autoprefixer';
 import cssnano from 'cssnano';
 import ExtractTextPlugin from 'extract-text-webpack-plugin';
 import minimist from 'minimist';
+import svgoConfig from './svgo.config';
 const argv = minimist(process.argv.slice(2));
 const NODE_ENV = process.env.NODE_ENV;
 const useRem = !!argv['use-rem'];
-const fileSuffix = useRem ? '' : '.px';
+const remSuffix = useRem ? '' : '.px';
+const libraryTarget = argv.libraryTarget || 'commonjs2';
+const libraryTargetSuffix = libraryTarget === 'umd' ? '.umd' : '';
+const fileSuffix = `${remSuffix}${libraryTargetSuffix}`;
 
 const entry = {
   bundle: './src/index.js',
@@ -17,13 +21,16 @@ const entry = {
     button: './src/components/button/button.jsx',
     'date-picker': './src/components/date-picker',
     dropdown: './src/components/dropdown/dropdown.jsx',
+    flag: ['./src/components/icon/icon.jsx'],
     graph: './src/components/graph',
     'graph-tooltip': './src/components/graph-tooltip',
     icon: ['./src/components/icon/icon.jsx'],
     input: './src/components/input/input.jsx',
     'labeled-value': './src/components/labeled-value/labeled-value.jsx',
+    legend: './src/components/legend',
     logo: './src/components/logo/logo.jsx',
     pane: './src/components/pane/pane.jsx',
+    'range-selector': './src/components/range-selector/',
     'ratio-bar': './src/components/ratio-bar/',
     select: './src/components/select/select.jsx',
     'spark-graph': ['./src/components/spark-graph'],
@@ -66,9 +73,22 @@ module.exports = (() => {
       path: './dist',
       filename: `${output[argv.type].filename}.js`,
       library: output[argv.type].library,
-      libraryTarget: 'commonjs2',
+      libraryTarget,
     },
-    externals: ['react', 'react-dom'],
+    externals: [{
+      react: {
+        root: 'React',
+        commonjs2: 'react',
+        commonjs: 'react',
+        amd: 'react',
+      },
+      'react-dom': {
+        root: 'ReactDOM',
+        commonjs2: 'react-dom',
+        commonjs: 'react-dom',
+        amd: 'react-dom',
+      },
+    }],
     resolve: {
       extensions: [
         '',
@@ -121,14 +141,6 @@ module.exports = (() => {
   // ////// //
   // Images //
   // ////// //
-
-  const svgoConfig = JSON.stringify({
-    plugins: [
-      { removeTitle: true },
-      { removeDimensions: true },
-      { convertColors: { shorthex: true } },
-    ],
-  });
 
   config.loader('svg', {
     test: /\.svg$/,